+++ src/main/java/com/brand/adapaxels/tags/APItemTags.java
@@ -1,18 +1,37 @@
 package com.brand.adapaxels.tags;
 
 import com.brand.adapaxels.AdaPaxels;
-import net.minecraft.item.Item;
-import net.minecraft.registry.RegistryKeys;
-import net.minecraft.registry.tag.TagKey;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.item.Item;
 
 public class APItemTags {
     public static final TagKey<Item> PAXELS = register("paxels");
+
+    public static final TagKey<Item> VIBRANIUM_TOOL_MATERIALS = register("vibranium_tool_materials");
+    public static final TagKey<Item> ADAMANTIUM_TOOL_MATERIALS = register("adamantium_tool_materials");
+    public static final TagKey<Item> NETHER_BRICK_TOOL_MATERIALS = register("nether_brick_tool_materials");
+
+    public static final TagKey<Item> TIN_TOOL_MATERIALS = register("tin_tool_materials");
+    public static final TagKey<Item> STEEL_TOOL_MATERIALS = register("steel_tool_materials");
+    public static final TagKey<Item> SILVER_TOOL_MATERIALS = register("silver_tool_materials");
+    public static final TagKey<Item> LEAD_TOOL_MATERIALS = register("lead_tool_materials");
+    public static final TagKey<Item> BRONZE_TOOL_MATERIALS = register("bronze_tool_materials");
+
+
+    public static final TagKey<Item> RUBY_TOOL_MATERIALS = register("ruby_tool_materials");
+    public static final TagKey<Item> SAPPHIRE_TOOL_MATERIALS = register("sapphire_tool_materials");
+    public static final TagKey<Item> PERIDOT_TOOL_MATERIALS = register("peridot_tool_materials");
+
+    public static final TagKey<Item> ENDERITE_TOOL_MATERIALS = register("enderite_tool_materials");
+
+    public static final TagKey<Item> DRAGON_TOOL_MATERIALS = register("dragon_tool_materials");
 
     private APItemTags() {
     }
 
     private static TagKey<Item> register(String id) {
-        return TagKey.of(RegistryKeys.ITEM, AdaPaxels.id(id));
+        return TagKey.create(Registries.ITEM, AdaPaxels.id(id));
     }
 }